♻️ [api] Move createAccessToken to pole emploi authentication service
