+++ api/lib/application/pole-emploi/pole-emploi-controller.js
@@ -1,5 +1,4 @@
 const usecases = require('../../domain/usecases');
-const tokenService = require('../../domain/services/token-service');
 const userRepository = require('../../infrastructure/repositories/user-repository');
 const poleEmploiAuthenticationService = require('../../../lib/domain/services/authentication/pole-emploi-authentication-service');
 
@@ -9,7 +8,7 @@ module.exports = {
 
     const { userId, idToken } = await usecases.createUserFromPoleEmploi({ authenticationKey });
 
-    const accessToken = tokenService.createAccessTokenForPoleEmploi(userId);
+    const accessToken = poleEmploiAuthenticationService.createAccessToken(userId);
     await userRepository.updateLastLoggedAt({ userId });
 
     const response = {
