+++ api/lib/domain/services/token-service.js
@@ -19,11 +19,6 @@ function createAccessTokenFromUser(userId, source) {
   return { accessToken, expirationDelaySeconds };
 }
 
-function createAccessTokenForPoleEmploi(userId) {
-  const expirationDelaySeconds = settings.poleEmploi.accessTokenLifespanMs / 1000;
-  return _createAccessToken({ userId, source: 'pole_emploi_connect', expirationDelaySeconds });
-}
-
 function createAccessTokenForSaml(userId) {
   const expirationDelaySeconds = settings.saml.accessTokenLifespanMs / 1000;
   return _createAccessToken({ userId, source: 'external', expirationDelaySeconds });
@@ -185,7 +180,6 @@ async function extractExternalUserFromIdToken(token) {
 
 module.exports = {
   createAccessTokenFromUser,
-  createAccessTokenForPoleEmploi,
   createAccessTokenForSaml,
   createAccessTokenFromApplication,
   createTokenForCampaignResults,
