+++ api/tests/unit/application/pole-emploi/pole-emploi-controller_test.js
@@ -3,7 +3,6 @@ const { expect, sinon, hFake } = require('../../../test-helper');
 const poleEmploiController = require('../../../../lib/application/pole-emploi/pole-emploi-controller');
 const usecases = require('../../../../lib/domain/usecases');
 const userRepository = require('../../../../lib/infrastructure/repositories/user-repository');
-const tokenService = require('../../../../lib/domain/services/token-service');
 const poleEmploiAuthenticationService = require('../../../../lib/domain/services/authentication/pole-emploi-authentication-service');
 
 describe('Unit | Controller | pole-emploi-controller', function () {
@@ -66,7 +65,7 @@ describe('Unit | Controller | pole-emploi-controller', function () {
       const request = { query: { 'authentication-key': 'abcde' } };
       const userId = 7;
       sinon.stub(usecases, 'createUserFromPoleEmploi').resolves({ userId, idToken: 1 });
-      sinon.stub(tokenService, 'createAccessTokenForPoleEmploi').resolves('an access token');
+      sinon.stub(poleEmploiAuthenticationService, 'createAccessToken').resolves('an access token');
       sinon.stub(userRepository, 'updateLastLoggedAt');
 
       // when
@@ -87,7 +86,7 @@ describe('Unit | Controller | pole-emploi-controller', function () {
         .withArgs({ authenticationKey: 'abcde' })
         .resolves({ userId, idToken });
       sinon.stub(userRepository, 'updateLastLoggedAt');
-      sinon.stub(tokenService, 'createAccessTokenForPoleEmploi').withArgs(userId).returns(accessToken);
+      sinon.stub(poleEmploiAuthenticationService, 'createAccessToken').withArgs(userId).returns(accessToken);
 
       // when
       const result = await poleEmploiController.createUser(request, hFake);
