+++ api/tests/unit/domain/services/authentication/pole-emploi-authentication-service_test.js
@@ -6,8 +6,28 @@ const httpAgent = require('../../../../../lib/infrastructure/http/http-agent');
 const poleEmploiAuthenticationService = require('../../../../../lib/domain/services/authentication/pole-emploi-authentication-service');
 const PoleEmploiTokens = require('../../../../../lib/domain/models/PoleEmploiTokens');
 const jsonwebtoken = require('jsonwebtoken');
+const { POLE_EMPLOI } = require('../../../../../lib/domain/constants').SOURCE;
 
 describe('Unit | Domain | Services | pole-emploi-authentication-service', function () {
+  describe('#createAccessToken', function () {
+    it('should create access token with user id and source', function () {
+      // given
+      const userId = 123;
+      settings.authentication.secret = 'a secret';
+      settings.poleEmploi.accessTokenLifespanMs = 1000;
+      const accessToken = 'valid access token';
+      const firstParameter = { user_id: userId, source: POLE_EMPLOI };
+      const secondParameter = 'a secret';
+      const thirdParameter = { expiresIn: 1 };
+      sinon.stub(jsonwebtoken, 'sign').withArgs(firstParameter, secondParameter, thirdParameter).returns(accessToken);
+
+      // when
+      const result = poleEmploiAuthenticationService.createAccessToken(userId);
+
+      // then
+      expect(result).to.equal(accessToken);
+    });
+  });
   describe('#exchangeCodeForTokens', function () {
     beforeEach(function () {
       sinon.stub(httpAgent, 'post');
